Changing to using char Arrays
After talking to Conner and Nick about problems with my code, I'm changing to from String arrays to Char arrays, as it will make everything easier. Doing so this late will cause my lab to be completed late, but I'd rather turn it in correct and late than on time and wrong.